# cursillo/swagger/dist/schema/swagger.json
{
  "swagger":"2.0",
  "info":{
    "description":"This is the Cursillo Swagger documentation.",
    "version":"1.0.0",
    "title":"Cursillo Swagger "
  },
  "host":"localhost",
  "basePath":":8080/api/",
...
  "schemes":[
    "http"
  ],
  "paths":{
...
    "/weekends/{weekendId}/talkLinks": {
      "get": {
        "tags": [
          "weekend"
        ],
        "summary": "return Talk Links for weekend with id weekendId.",
        "description": "",
        "operationId": "getTalkLinksByWeekendId",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "authorization",
            "description": "access token",
            "required": true,
            "type": "string"
          },
          {
            "in": "path",
            "name": "weekendId",
            "description": "weekend id",
            "required": true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses": {
          "400": {
            "description": "invalid credentials."
          }
        },
        "security": [
        ]
      }
    },
...
    "/talks": {
      "get":{
        "tags":[
          "talk"
        ],
        "summary":"Gets all talks.",
        "description":"",
        "operationId":"Get Talks",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "post":{
        "tags":[
          "talk"
        ],
        "summary":"Creates a new talk.",
        "description":"",
        "operationId":"createTalk",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"body",
            "name":"body",
            "description":"Create a new talk",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Talk"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
    "/talks/{talkId}": {
      "get":{
        "tags":[
          "talk"
        ],
        "summary":"Return talk with Id talkId.",
        "description":"",
        "operationId":"getTalkbyId",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"talkId",
            "description":"Talk Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "put":{
        "tags":[
          "talk"
        ],
        "summary":"Updates an existing Talk.",
        "description":"",
        "operationId":"updateTalk",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"talkId",
            "description":"Id of talk being updated.",
            "required":true
          },
          {
            "in":"body",
            "name":"body",
            "description":"Create a new talk.",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Talk"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "delete":{
        "tags":[
          "talk"
        ],
        "summary":"Delete Talk with Id talkId.",
        "description":"",
        "operationId":"deleteTalkById",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"talkId",
            "description":"Talk Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
    "/talks/{talkId}/people": {
      "get":{
        "tags":[
          "talk"
        ],
        "summary":"Return people that have had talk with Id talkId.",
        "description":"",
        "operationId":"getPeopleByTalkId",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"talkId",
            "description":"Talk Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
    "/talkLinks": {
      "get":{
        "tags":[
          "talkLink"
        ],
        "summary":"Gets all talkLinks.",
        "description":"",
        "operationId":"Get TalkLinks",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "post":{
        "tags":[
          "talkLink"
        ],
        "summary":"Creates a new talkLink.",
        "description":"",
        "operationId":"createTalkLink",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"body",
            "name":"body",
            "description":"Create a new talkLink",
            "required":true,
            "schema":{
              "$ref":"#/definitions/TalkLink"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
    "/talkLinks/{talkLinkId}": {
      "get":{
        "tags":[
          "talkLink"
        ],
        "summary":"Return talkLink with Id talkLinkId.",
        "description":"",
        "operationId":"getTalkLinkbyId",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"talkLinkId",
            "description":"TalkLink Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "put":{
        "tags":[
          "talkLink"
        ],
        "summary":"Updates an existing TalkLink.",
        "description":"",
        "operationId":"updateTalkLink",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"talkLinkId",
            "description":"Id of talkLink being updated.",
            "required":true
          },
          {
            "in":"body",
            "name":"body",
            "description":"Create a new talkLink.",
            "required":true,
            "schema":{
              "$ref":"#/definitions/TalkLink"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "delete":{
        "tags":[
          "talkLink"
        ],
        "summary":"Delete TalkLink with Id talkLinkId.",
        "description":"",
        "operationId":"deleteTalkLinkById",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"talkLinkId",
            "description":"TalkLink Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
    "/people": {
      "get":{
        "tags":[
          "person"
        ],
        "summary":"Gets all people.",
        "description":"",
        "operationId":"Get People",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      },
      "post":{
        "tags":[
          "person"
        ],
        "summary":"Creates a new person.",
        "description":"",
        "operationId":"createPerson",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"body",
            "name":"body",
            "description":"Create a new person",
            "required":true,
            "schema":{
              "$ref":"#/definitions/Person"
            }
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
...
    "/people/{personId}/talkLinks": {
      "get":{
        "tags":[
          "person"
        ],
        "summary":"Return person with Id personId.",
        "description":"",
        "operationId":"getTalkLinksByPersonId",
        "consumes":[
          "application/json"
        ],
        "produces":[
          "application/json"
        ],
        "parameters":[
          {
            "in":"query",
            "name":"authorization",
            "description":"access token",
            "required":true,
            "type": "string"
          },
          {
            "in":"path",
            "name":"personId",
            "description":"Person Id",
            "required":true,
            "type": "integer",
            "form": "int32"
          }
        ],
        "responses":{
          "400":{
            "description":"invalid credentials."
          }
        },
        "security":[

        ]
      }
    },
...
  },
  "securityDefinitions":{
    "petstore_auth":{
      "type":"oauth2",
      "authorizationUrl":"http://petstore.swagger.io/oauth/dialog",
      "flow":"implicit",
      "scopes":{
        "write:pets":"modify pets in your account",
        "read:pets":"read your pets"
      }
    },
    "api_key":{
      "type":"apiKey",
      "name":"api_key",
      "in":"header"
    }
  },
  "definitions":{
...
    "Talk": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "required": "true"
        },
        "name": {
          "type": "string",
          "required": "true"
        },
        "number": {
          "type": "number"
        },
        "description": {
          "type": "string"
        },
        "isActive": {
          "type": "boolean"
        }
      }
    },
    "TalkLink": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "required": "true"
        },
        "status": {
          "type": "string",
          "required": "true"
        },
        "talkId": {
          "type": "integer",
          "format": "int64",
          "required": true
        },
        "personId": {
          "type": "integer",
          "format": "int64",
          "required": true
        },
        "weekendId": {
          "type": "integer",
          "format": "int64",
          "required": true
        }

      }
    },
    "Person": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64",
          "required": true
        },
        "firstName": {
          "type": "string",
          "required": true
        },
        "lastName": {
          "type": "string",
          "required": true
        },
        "email": {
          "type": "string",
          "required": true
        },
        "homePhone": {
          "type": "string"
        },
        "workPhone": {
          "type": "string"
        },
        "street": {
          "type": "string"
        },
        "city": {
          "type": "string"
        },
        "state": {
          "type": "string"
        },
        "zip": {
          "type": "number"
        },
        "birthdate": {
          "type": "string"
        },
        "occupation": {
          "type": "string"
        },
        "gender": {
          "type": "string"
        },
        "isTeamMember": {
          "type": "boolean",
          "required": true,
          "default": false
        },
        "dateStarted": {
          "type": "string"
        },
        "tagName": {
          "type": "string"
        },
        "parishId": {
          "type": "integer",
          "format": "int64",
          "required": true
        }
      }
    },
...
  }
}